Fixes to test framework and GitHub workflow
After many tries, I think I finally manage to find the right combination of
tweaks to make pytest, tox, and GitHub actions to work.

This also adds caching of the librocksdb compilation, which is a very slow
step, and restricts building of wheels and other artifacts to run only when
pushing to main.